tttt

+++ webCrawler/src/main/java/com/company/IndexerObject.java
@@ -5,6 +5,6 @@ import java.util.ArrayList;
 public class IndexerObject {
     public int TF = 0;
     public int Weight = 0;
-    public int paragraphID = -1;
+    public long paragraphID = -1;
     public ArrayList<Integer> positions = new ArrayList<>();
 }
